Prepared actors hierarchy for migration.
Internal nodes added so methods relevant to akka can be overridden.

Review by: @phaller
